chore(nextjs): Drop `withClerkMiddleware`
