+++ packages/nextjs/src/server-helpers.client.ts
@@ -7,6 +7,5 @@ export const currentUser = () => mockImplementation('currentUser()');
 export const authMiddleware = () => mockImplementation('authMiddleware()');
 export const getAuth = () => mockImplementation('getAuth()');
 export const clerkClient = () => mockImplementation('clerkClient()');
-export const withClerkMiddleware = () => mockImplementation('withClerkMiddleware()');
 export const redirectToSignIn = () => mockImplementation('redirectToSignIn()');
 export const redirectToSignUp = () => mockImplementation('redirectToSignUp()');
