+++ packages/nextjs/src/server-helpers.server.ts
@@ -4,18 +4,8 @@ import { authMiddleware } from './server/authMiddleware';
 import { clerkClient } from './server/clerkClient';
 import { getAuth } from './server/getAuth';
 import { redirectToSignIn, redirectToSignUp } from './server/redirect';
-import { withClerkMiddleware } from './server/withClerkMiddleware';
 
-export {
-  auth,
-  currentUser,
-  authMiddleware,
-  getAuth,
-  clerkClient,
-  withClerkMiddleware,
-  redirectToSignIn,
-  redirectToSignUp,
-};
+export { auth, authMiddleware, clerkClient, currentUser, getAuth, redirectToSignIn, redirectToSignUp };
 
 export type ServerHelpersServerModuleTypes = {
   auth: typeof auth;
@@ -23,7 +13,6 @@ export type ServerHelpersServerModuleTypes = {
   authMiddleware: typeof authMiddleware;
   getAuth: typeof getAuth;
   clerkClient: typeof clerkClient;
-  withClerkMiddleware: typeof withClerkMiddleware;
   redirectToSignIn: typeof redirectToSignIn;
   redirectToSignUp: typeof redirectToSignUp;
 };
