+++ packages/nextjs/src/server/authMiddleware.test.ts
@@ -1,5 +1,6 @@
 // There is no need to execute the complete authenticateRequest to test authMiddleware
 // This mock SHOULD exist before the import of authenticateRequest
+import { expectTypeOf } from 'expect-type';
 import { NextURL } from 'next/dist/server/web/next-url';
 import type { NextFetchEvent, NextRequest } from 'next/server';
 import { NextResponse } from 'next/server';
@@ -596,5 +597,60 @@ describe('401 Response on Api Routes', function () {
     })(mockRequest({ url: '/public' }), {} as NextFetchEvent);
 
     expect(resp?.status).toEqual(200);
+  });
+});
+
+describe('Type tests', () => {
+  type AuthMiddleware = Parameters<typeof authMiddleware>[0];
+  describe('AuthMiddleware', () => {
+    it('is the options argument for authMiddleware', () => {
+      () => {
+        authMiddleware({} as AuthMiddleware);
+      };
+    });
+
+    it('can receive the appropriate keys', () => {
+      expectTypeOf({ publishableKey: '', secretKey: '' }).toMatchTypeOf<AuthMiddleware>();
+      expectTypeOf({ secretKey: '' }).toMatchTypeOf<AuthMiddleware>();
+      expectTypeOf({ publishableKey: '', secretKey: '' }).toMatchTypeOf<AuthMiddleware>();
+      expectTypeOf({ secretKey: '' }).toMatchTypeOf<AuthMiddleware>();
+    });
+
+    describe('Multi domain', () => {
+      const defaultProps = { publishableKey: '', secretKey: '' };
+
+      it('proxyUrl (primary app)', () => {
+        expectTypeOf({ ...defaultProps, proxyUrl: 'test' }).toMatchTypeOf<AuthMiddleware>();
+      });
+
+      it('proxyUrl + isSatellite (satellite app)', () => {
+        expectTypeOf({ ...defaultProps, proxyUrl: 'test', isSatellite: true }).toMatchTypeOf<AuthMiddleware>();
+      });
+
+      it('domain + isSatellite (satellite app)', () => {
+        expectTypeOf({ ...defaultProps, domain: 'test', isSatellite: true }).toMatchTypeOf<AuthMiddleware>();
+      });
+
+      it('only domain is not allowed', () => {
+        expectTypeOf({ ...defaultProps, domain: 'test' }).not.toMatchTypeOf<AuthMiddleware>();
+      });
+
+      it('only isSatellite is not allowed', () => {
+        expectTypeOf({ ...defaultProps, isSatellite: true }).not.toMatchTypeOf<AuthMiddleware>();
+      });
+
+      it('proxyUrl + domain is not allowed', () => {
+        expectTypeOf({ ...defaultProps, proxyUrl: 'test', domain: 'test' }).not.toMatchTypeOf<AuthMiddleware>();
+      });
+
+      it('proxyUrl + domain + isSatellite is not allowed', () => {
+        expectTypeOf({
+          ...defaultProps,
+          proxyUrl: 'test',
+          domain: 'test',
+          isSatellite: true,
+        }).not.toMatchTypeOf<AuthMiddleware>();
+      });
+    });
   });
 });
