+++ packages/nextjs/src/server/authenticateRequest.ts
@@ -1,18 +1,11 @@
+import { constants, debugRequestState } from '@clerk/backend';
 import type { NextRequest } from 'next/server';
 import { NextResponse } from 'next/server';
 
 import type { RequestState } from './clerkClient';
-import {
-  CLERK_JS_URL,
-  CLERK_JS_VERSION,
-  clerkClient,
-  debugRequestState,
-  PUBLISHABLE_KEY,
-  SECRET_KEY,
-} from './clerkClient';
+import { CLERK_JS_URL, CLERK_JS_VERSION, clerkClient, PUBLISHABLE_KEY, SECRET_KEY } from './clerkClient';
 import type { WithAuthOptions } from './types';
 import { apiEndpointUnauthorizedNextResponse, handleMultiDomainAndProxy } from './utils';
-import { decorateResponseWithObservabilityHeaders } from './withClerkMiddleware';
 
 export const authenticateRequest = async (req: NextRequest, opts: WithAuthOptions) => {
   const { isSatellite, domain, signInUrl, proxyUrl } = handleMultiDomainAndProxy(req, opts);
@@ -26,6 +19,12 @@ export const authenticateRequest = async (req: NextRequest, opts: WithAuthOption
     proxyUrl,
     request: req,
   });
+};
+
+const decorateResponseWithObservabilityHeaders = (res: NextResponse, requestState: RequestState) => {
+  requestState.message && res.headers.set(constants.Headers.AuthMessage, encodeURIComponent(requestState.message));
+  requestState.reason && res.headers.set(constants.Headers.AuthReason, encodeURIComponent(requestState.reason));
+  requestState.status && res.headers.set(constants.Headers.AuthStatus, encodeURIComponent(requestState.status));
 };
 
 export const handleUnknownState = (requestState: RequestState) => {
