+++ packages/nextjs/src/server/index.ts
@@ -1,7 +1,6 @@
 export * from './clerkClient';
 export { buildClerkProps, getAuth } from './getAuth';
-export * from './withClerkMiddleware';
-export { redirectToSignUp, redirectToSignIn } from './redirect';
+export { redirectToSignIn, redirectToSignUp } from './redirect';
 
 export { auth } from '../app-router/server/auth';
 export { currentUser } from '../app-router/server/currentUser';
